fix: rm coc-tabnine extension
was causing very high cpu load and memory, creating a laggy experience. Tried to log into the reasoning, appears a common issue as per following:

https://github.com/zxqfl/TabNine/issues/215
https://github.com/zxqfl/TabNine/issues/201
https://github.com/zxqfl/TabNine/issues/183
and on and on...

and to my surprise, of course it is closed-source:

https://github.com/zxqfl/TabNine/issues/4
https://github.com/zxqfl/TabNine/issues/179

so can only presume that magic backend is some terrible spaghetti code in the best case, or has the potential for malicious actions in the worst case. No thanks!